Upgrade package to TypeScript 🎆
