+++ src/.d.ts
@@ -1,0 +1,4 @@
+declare module 'babili-webpack-plugin'
+declare module 'clean-webpack-plugin'
+declare module 'duplicate-package-checker-webpack-plugin'
+declare module 'webpack-visualizer-plugin'
